requirements.txt
flask>=3.0
flask-cors>=4.0
pyautogui>=0.9.54
pyperclip>=1.8
pillow>=10.0
python-dotenv>=1.0
requests>=2.31
discord.py>=2.3
aiohttp>=3.9
